Renamed Translator classes within dom-translation scripts to DomTranslator equivalents to clarify what they are
